Make only one sync request at a time
Prevent situations when when two sync requests are sent, and the second one
returns earlier than the first one, and the first one fails.
This leads to data inconsistency: since client sync time will be set
to the time when the second request's response is received, the first
chunk of data will be treated as synchronized whereas in reality that
particular sync failed.