Introduce creationDate property
Using this new property instead of createdAt, we ensure that
creation dates associated with flashcards stay the same regardless
of time zone.
Thus, even though the client can be in any time zone, once a flashcard
is created, its creation date stays the same on the server and all clients.